Add separate dynamic catergory and add new category files
